Make package compatible python2.7 and add LICENSE

+++ searchlight/utils.py
@@ -6,4 +6,4 @@ EPOCH = datetime.datetime.strptime("07-26-2009", "%m-%d-%Y")
 
 
 def week_number(date):
-    return math.ceil((datetime.datetime.strptime(date, "%m-%d-%Y") - EPOCH).days / 7)
+    return int(math.ceil((datetime.datetime.strptime(date, "%m-%d-%Y") - EPOCH).days / 7))
